Move getValue/setValue to library functions
Previously these two runtime functions lived in `runtime_safe_heap.js`.
This change makes them into library functions and splits other parts of
`runtime_safe_heap.js` into `runtime_asan.js` so that
`runtime_safe_heap.js` not only contains code that is included/needed
when SAFE_HEAP is enabled.

One of the main effects of this change is that getValue/setValue are not
potentially available in MINIMAL_RUNTIME, although I've left using them
this mode as a followup.
