+++ src/preamble_minimal.js
@@ -4,7 +4,13 @@
  * SPDX-License-Identifier: MIT
  */
 
+#if SAFE_HEAP
 #include "runtime_safe_heap.js"
+#endif
+
+#if USE_ASAN
+#include "runtime_asan.js"
+#endif
 
 #if ASSERTIONS || SAFE_HEAP
 /** @type {function(*, string=)} */
