+++ src/runtime_safe_heap.js
@@ -4,82 +4,10 @@
  * SPDX-License-Identifier: MIT
  */
 
-#if SAFE_HEAP || !MINIMAL_RUNTIME
-// In MINIMAL_RUNTIME, setValue() and getValue() are only available when
-// building with safe heap enabled, for heap safety checking.
-// In traditional runtime, setValue() and getValue() are always available
-// (although their use is highly discouraged due to perf penalties)
-
-/** @param {number} ptr
-    @param {number} value
-    @param {string} type
-    @param {number|boolean=} noSafe */
-function setValue(ptr, value, type = 'i8', noSafe) {
-  if (type.endsWith('*')) type = '{{{ POINTER_TYPE }}}';
-#if SAFE_HEAP
-  if (noSafe) {
-    switch (type) {
-      case 'i1': {{{ makeSetValue('ptr', '0', 'value', 'i1', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      case 'i8': {{{ makeSetValue('ptr', '0', 'value', 'i8', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      case 'i16': {{{ makeSetValue('ptr', '0', 'value', 'i16', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      case 'i32': {{{ makeSetValue('ptr', '0', 'value', 'i32', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      case 'i64': {{{ makeSetValue('ptr', '0', 'value', 'i64', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      case 'float': {{{ makeSetValue('ptr', '0', 'value', 'float', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      case 'double': {{{ makeSetValue('ptr', '0', 'value', 'double', undefined, undefined, undefined, /*noSafe=*/true) }}}; break;
-      default: abort('invalid type for setValue: ' + type);
-    }
-    // unreachable
-    return;
-  }
+#if !SAFE_HEAP
+#error "should only be inclded in SAFE_HEAP mode"
 #endif
-  switch (type) {
-    case 'i1': {{{ makeSetValue('ptr', '0', 'value', 'i1') }}}; break;
-    case 'i8': {{{ makeSetValue('ptr', '0', 'value', 'i8') }}}; break;
-    case 'i16': {{{ makeSetValue('ptr', '0', 'value', 'i16') }}}; break;
-    case 'i32': {{{ makeSetValue('ptr', '0', 'value', 'i32') }}}; break;
-    case 'i64': {{{ makeSetValue('ptr', '0', 'value', 'i64') }}}; break;
-    case 'float': {{{ makeSetValue('ptr', '0', 'value', 'float') }}}; break;
-    case 'double': {{{ makeSetValue('ptr', '0', 'value', 'double') }}}; break;
-    default: abort('invalid type for setValue: ' + type);
-  }
-}
 
-/** @param {number} ptr
-    @param {string} type
-    @param {number|boolean=} noSafe */
-function getValue(ptr, type = 'i8', noSafe) {
-  if (type.endsWith('*')) type = '{{{ POINTER_TYPE }}}';
-#if SAFE_HEAP
-  if (noSafe) {
-    switch (type) {
-      case 'i1': return {{{ makeGetValue('ptr', '0', 'i1', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      case 'i8': return {{{ makeGetValue('ptr', '0', 'i8', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      case 'i16': return {{{ makeGetValue('ptr', '0', 'i16', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      case 'i32': return {{{ makeGetValue('ptr', '0', 'i32', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      case 'i64': return {{{ makeGetValue('ptr', '0', 'i64', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      case 'float': return {{{ makeGetValue('ptr', '0', 'float', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      case 'double': return {{{ makeGetValue('ptr', '0', 'double', undefined, undefined, undefined, undefined, /*noSafe=*/true) }}};
-      default: abort('invalid type for getValue: ' + type);
-    }
-    // unreachable
-    return;
-  }
-#endif
-  switch (type) {
-    case 'i1': return {{{ makeGetValue('ptr', '0', 'i1') }}};
-    case 'i8': return {{{ makeGetValue('ptr', '0', 'i8') }}};
-    case 'i16': return {{{ makeGetValue('ptr', '0', 'i16') }}};
-    case 'i32': return {{{ makeGetValue('ptr', '0', 'i32') }}};
-    case 'i64': return {{{ makeGetValue('ptr', '0', 'i64') }}};
-    case 'float': return {{{ makeGetValue('ptr', '0', 'float') }}};
-    case 'double': return {{{ makeGetValue('ptr', '0', 'double') }}};
-    default: abort('invalid type for getValue: ' + type);
-  }
-}
-#endif // SAFE_HEAP || !MINIMAL_RUNTIME
-
-
-#if SAFE_HEAP
 /** @param {number|boolean=} isFloat */
 function getSafeHeapType(bytes, isFloat) {
   switch (bytes) {
@@ -177,92 +105,3 @@ function alignfault() {
   warnOnce('alignment fault');
 #endif
 }
-#endif
-
-#if USE_ASAN
-// C versions of asan_js_{load|store}_* will be used from compiled code, which have
-// ASan instrumentation on them. However, until the wasm module is ready, we
-// must access things directly.
-
-/** @suppress{duplicate} */
-function _asan_js_load_1(ptr) {
-  if (runtimeInitialized) return _asan_c_load_1(ptr);
-  return HEAP8[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_1u(ptr) {
-  if (runtimeInitialized) return _asan_c_load_1u(ptr);
-  return HEAPU8[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_2(ptr) {
-  if (runtimeInitialized) return _asan_c_load_2(ptr);
-  return HEAP16[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_2u(ptr) {
-  if (runtimeInitialized) return _asan_c_load_2u(ptr);
-  return HEAPU16[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_4(ptr) {
-  if (runtimeInitialized) return _asan_c_load_4(ptr);
-  return HEAP32[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_4u(ptr) {
-  if (runtimeInitialized) return _asan_c_load_4u(ptr) >>> 0;
-  return HEAPU32[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_f(ptr) {
-  if (runtimeInitialized) return _asan_c_load_f(ptr);
-  return HEAPF32[ptr];
-}
-/** @suppress{duplicate} */
-function _asan_js_load_d(ptr) {
-  if (runtimeInitialized) return _asan_c_load_d(ptr);
-  return HEAPF64[ptr];
-}
-
-/** @suppress{duplicate} */
-function _asan_js_store_1(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_1(ptr, val);
-  return HEAP8[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_1u(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_1u(ptr, val);
-  return HEAPU8[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_2(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_2(ptr, val);
-  return HEAP16[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_2u(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_2u(ptr, val);
-  return HEAPU16[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_4(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_4(ptr, val);
-  return HEAP32[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_4u(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_4u(ptr, val) >>> 0;
-  return HEAPU32[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_f(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_f(ptr, val);
-  return HEAPF32[ptr] = val;
-}
-/** @suppress{duplicate} */
-function _asan_js_store_d(ptr, val) {
-  if (runtimeInitialized) return _asan_c_store_d(ptr, val);
-  return HEAPF64[ptr] = val;
-}
-#endif
